Support time range to query dimension names/values
At present, dimensions are not scoped by time window, which makes
dimension related queries to large databases timeout because it searches
all of time instead of a time window specified on the grafana app. This
commit aims to address issue by scoping the search to time range
specified on the app.

Change-Id: I13614595707697552119926273f06967b8d49b1d
Story: 2006204
Task: 35789